@@ -29,11 +29,12 @@
 
 package metis.dictionary;
 
+import com.formdev.flatlaf.FlatLaf;
+import com.formdev.flatlaf.FlatLightLaf;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
-import javax.swing.plaf.ColorUIResource;
-import javax.swing.plaf.metal.DefaultMetalTheme;
-import javax.swing.plaf.metal.MetalLookAndFeel;
 
 import metis.dictionary.ui.DictionaryFrame;
 
@@ -45,6 +46,12 @@ import metis.dictionary.ui.DictionaryFrame;
  */
 public class Dictionary {
 
+    private static final Logger LOG = Logger.getLogger(Dictionary.class.getName());
+
+    static {
+        System.setProperty("java.util.logging.SimpleFormatter.format", "[%1$tF %1$tT] [%4$-7s] %5$s %n");
+    }
+
     /**
      * Main method.
      *
@@ -52,57 +59,15 @@ public class Dictionary {
      */
     public static void main(final String[] args) {
 
-        MetalLookAndFeel.setCurrentTheme(new DefaultMetalTheme() {
-            private final ColorUIResource primary1 = new ColorUIResource(0x4d6781);
-            private final ColorUIResource primary2 = new ColorUIResource(0x7a96b0);
-            private final ColorUIResource primary3 = new ColorUIResource(0xc8d4e2);
-            private final ColorUIResource secondary1 = new ColorUIResource(0x000000);
-            private final ColorUIResource secondary2 = new ColorUIResource(0xaaaaaa);
-            private final ColorUIResource secondary3 = new ColorUIResource(0xdfdfdf);
-
-            @Override
-            protected ColorUIResource getPrimary1() {
-                return this.primary1;
-            }
-
-            @Override
-            protected ColorUIResource getPrimary2() {
-                return this.primary2;
-            }
-
-            @Override
-            protected ColorUIResource getPrimary3() {
-                return this.primary3;
-            }
-
-            @Override
-            protected ColorUIResource getSecondary1() {
-                return this.secondary1;
-            }
-
-            @Override
-            protected ColorUIResource getSecondary2() {
-                return this.secondary2;
-            }
-
-            @Override
-            protected ColorUIResource getSecondary3() {
-                return this.secondary3;
-            }
-        });
-
         try {
-            UIManager.put("swing.boldMetal", Boolean.FALSE);
-            UIManager.setLookAndFeel("javax.swing.plaf.metal.MetalLookAndFeel");
+            UIManager.put("Button.arc", 0);
+            UIManager.put("Component.arc", 0);
+            FlatLightLaf.install();
+            FlatLaf lookAndFeel = new FlatLightLaf();
+            UIManager.setLookAndFeel(lookAndFeel);
         } catch (Exception e) {
-            e.printStackTrace();
+            LOG.log(Level.CONFIG, "Could not set look and feel for the application", e);
         }
-
-        SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                DictionaryFrame.getInstance();
-            }
-        });
+        SwingUtilities.invokeLater(() -> DictionaryFrame.getInstance());
     }
 }
